chore: organise operator implementations for `Expression`
