+++ acvm/src/compiler/transformers/fallback.rs
@@ -70,8 +70,8 @@ impl FallbackTransformer {
             BlackBoxFunc::AND => {
                 let (lhs, rhs, result, num_bits) = crate::pwg::logic::extract_input_output(gc);
                 stdlib::fallback::and(
-                    Expression::from(&lhs),
-                    Expression::from(&rhs),
+                    Expression::from(lhs),
+                    Expression::from(rhs),
                     result,
                     num_bits,
                     current_witness_idx,
@@ -80,8 +80,8 @@ impl FallbackTransformer {
             BlackBoxFunc::XOR => {
                 let (lhs, rhs, result, num_bits) = crate::pwg::logic::extract_input_output(gc);
                 stdlib::fallback::xor(
-                    Expression::from(&lhs),
-                    Expression::from(&rhs),
+                    Expression::from(lhs),
+                    Expression::from(rhs),
                     result,
                     num_bits,
                     current_witness_idx,
@@ -93,7 +93,7 @@ impl FallbackTransformer {
                 let input = &gc.inputs[0];
                 // Note there are no outputs because range produces no outputs
                 stdlib::fallback::range(
-                    Expression::from(&input.witness),
+                    Expression::from(input.witness),
                     input.num_bits,
                     current_witness_idx,
                 )
